feat: more robust error handling

+++ transcoder/resumable.py
@@ -9,6 +9,7 @@ import datetime
 import io
 import os
 import sqlite3
+import sys
 
 from tqdm import tqdm
 
@@ -58,6 +59,7 @@ class ResumableFileSet:
     cur.execute("""DROP TABLE IF EXISTS filelist""")
     cur.execute("""DROP TABLE IF EXISTS xfermeta""")
     cur.execute("""DROP TABLE IF EXISTS stats""")
+    cur.execute("""DROP TABLE IF EXISTS errors""")
     cur.close()
 
   def create(
@@ -74,6 +76,7 @@ class ResumableFileSet:
     cur.execute("""DROP TABLE IF EXISTS filelist""")
     cur.execute("""DROP TABLE IF EXISTS xfermeta""")
     cur.execute("""DROP TABLE IF EXISTS stats""")
+    cur.execute("""DROP TABLE IF EXISTS errors""")
 
     cur.execute(f"""
       CREATE TABLE xfermeta (
@@ -89,8 +92,10 @@ class ResumableFileSet:
     """)
 
     cur.execute(
-      "INSERT INTO xfermeta VALUES (?,?,?,?,?,?,?,?)", 
-      [ 1, src, dest, recompress, reencode, delete_original, level, now_msec() ]
+      """INSERT INTO xfermeta 
+      (id, source, dest, recompress, reencode, encoding_level, delete_original, created) 
+      VALUES (?,?,?,?,?,?,?,?)""", 
+      [ 1, src, dest, recompress, reencode, level, delete_original, now_msec() ]
     )
 
     cur.execute(f"""
@@ -126,6 +131,13 @@ class ResumableFileSet:
     )
 
     cur.close()
+
+  def errors(self, n=1000):
+    cur = self.conn.cursor()
+    cur.execute(f"SELECT filename, error, created FROM errors LIMIT {int(n)}")
+    results = cur.fetchmany()
+    cur.close()
+    return results
 
   def record_error(self, filename, error):
     cur = self.conn.cursor()
@@ -222,7 +234,7 @@ class ResumableFileSet:
     cur.close()
     return int(res[0])
 
-  def total(self):
+  def total(self) -> int:
     """Returns the total number of tasks (both processed and unprocessed)."""
     if not self._total_dirty:
       return self._total
@@ -231,19 +243,25 @@ class ResumableFileSet:
     self._total_dirty = False
     return self._total
 
-  def finished(self):
+  def finished(self) -> int:
     return self._scalar_query(f"SELECT value FROM stats WHERE id = 1")
 
-  def remaining(self):
+  def remaining(self) -> int:
     return self.total() - self.finished()
 
-  def num_leased(self):
+  def num_leased(self) -> int:
     ts = int(now_msec())
     return self._scalar_query(
       f"SELECT count(filename) FROM filelist WHERE finished = 0 AND lease > {ts}"
     )
 
-  def available(self):
+  def num_errors(self) -> int:
+    return self._scalar_query(f"SELECT count(*) from errors")
+
+  def has_errors(self) -> bool:
+    return self._scalar_query(f"SELECT count(*) from errors limit 1") > 0
+
+  def available(self) -> int:
     ts = int(now_msec())
     return self._scalar_query(
       f"SELECT count(filename) FROM filelist WHERE finished = 0 AND lease <= {ts}"
@@ -357,6 +375,10 @@ class ResumableTransfer:
         pbar.refresh()
 
   def close(self):
+    if self.rfs.has_errors():
+      print("There were errors during processing. Keeping the database intact.", file=sys.stderr)
+      return
+
     self.rfs.delete()
     try:
       os.remove(self.db_path)
